Changing BaudRate back to 9600
Now, BaudRate is the rate at which the signal is sampled from the serial
port. Decreasing it slowed the rate at which the graph was generated
(which was why we looked silly holding onto a solid-state thermometer
for ~10 seconds). I'm changing it back to 9600 for reasons. More
specifically, those reasons are mostly due to accuracy (I think), and
everywhere else has it at 9600 Baums.